# Add `NewFromByteString` function