--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -20,7 +20,7 @@ jobs:
     strategy:
       max-parallel: 2
       matrix:
-        version: ['1.19', '1.20', '1.21', '1.22']
+        version: ['1.19', '1.20', '1.21', '1.22', '1.23', '1.24']
 
     name: Go ${{ matrix.version }}
     steps: