@PauliusLozys 2025-03-20
Tweaked `NewFromString` and `NewFromByteString` functions to do a single loop check for dots and scientific notations. This added additional improvements of about 10%, bringing total to ~40%
Before
```
goos: darwin
goarch: arm64
pkg: github.com/advbet/xdecimal/v2
cpu: Apple M1 Pro
BenchmarkDecimal_UnmarshalJSON-8         9977294               111.4 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10686229               112.1 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10558333               112.3 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10628580               112.9 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10602610               112.0 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10629855               111.1 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10584810               111.4 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10594178               111.5 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10626913               112.3 ns/op            56 B/op          4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        10498070               114.7 ns/op            56 B/op          4 allocs/op
PASS
ok      github.com/advbet/xdecimal/v2   13.259s
```

After 
```
goos: darwin
goarch: arm64
pkg: github.com/advbet/xdecimal/v2
cpu: Apple M1 Pro
BenchmarkDecimal_UnmarshalJSON-8        19228134                63.62 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        18327243                64.86 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        18148636                64.53 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        18836480                65.39 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        18307718                64.58 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        18033180                63.70 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        18892318                62.33 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        19206991                61.97 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        18704613                61.83 ns/op           48 B/op          3 allocs/op
BenchmarkDecimal_UnmarshalJSON-8        19247760                61.65 ns/op           48 B/op          3 allocs/op
PASS
ok      github.com/advbet/xdecimal/v2   13.735s
```

Comparison:
```
goos: darwin
goarch: arm64
pkg: github.com/advbet/xdecimal/v2
cpu: Apple M1 Pro
                        │  bench1.txt  │             bench2.txt              │
                        │    sec/op    │   sec/op     vs base                │
Decimal_UnmarshalJSON-8   112.05n ± 1%   63.66n ± 3%  -43.19% (p=0.000 n=10)

                        │ bench1.txt │             bench2.txt             │
                        │    B/op    │    B/op     vs base                │
Decimal_UnmarshalJSON-8   56.00 ± 0%   48.00 ± 0%  -14.29% (p=0.000 n=10)

                        │ bench1.txt │             bench2.txt             │
                        │ allocs/op  │ allocs/op   vs base                │
Decimal_UnmarshalJSON-8   4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.000 n=10)
```